added zotbib-rofi script and helper
opening the zotero application everytime felt cumbersome especially
when i knew exactly what paper i needed to read. so this little script
and its big helper script. the helper script (in python) parses the bib
file that is automatically updated on each zotero usage, and outputs
string that can be used by rofi. the helper also maintains a databse
of pdf file location in order with the rofi prompt. the actual script
(in bash) uses the output from helper, passes it through rofi, and uses
the index to open the pdf file from the database. as simple as that.